v0.2.0 - Maybe, possibly, actually this time Fixed take all Added selling items at a mark up Blacksmith now sells iron bars Added the woods Moved some campfire recipes to the grove Added traders Can now sell to the blacksmith Added Mysterious Trader You can now drop items Added XP share Buffed some enemies from the last update Low-level enemies spawn less on lower floors Dungeon min floors increased: 5 -> 7 Dungeon max floors increased: 7 -> 9 Added teleporation scrolls Added bank Added vaults Fixed weapon switching not updating attacks issue Logs to coal no longer gives XP Consumables are actually consumed now Fixed directions Added a line break before CON in the level up menu Added a forsaken place Players can now start the beginning of the end 8 new monsters 17 new items 3 new dungeon objects

+++ server/src/ItemTypes/Weapon.cs
@@ -18,7 +18,7 @@ namespace ItemTypes
 
         protected override bool EquipInHands => true;
 
-        public Weapon(string id, string name, Die damage, DamageType damageType, string description = "No description provided", int weight = 0, int sellValue = 0, 
+        public Weapon(string id, string name, Die damage, DamageType damageType, string description = "No description provided", float weight = 0, int sellValue = 0, 
             AbilityScore? abilityScore = null, string color = "white")
             : base(id, name, weight, description, color)
         {
@@ -27,7 +27,7 @@ namespace ItemTypes
             this.sellValue = sellValue;
         }
 
-        public Weapon(string id, string name, Attack[] attacks, string description = "No description provided", int weight = 0, int sellValue = 0, string color = "white")
+        public Weapon(string id, string name, Attack[] attacks, string description = "No description provided", float weight = 0, int sellValue = 0, string color = "white")
             : base(id, name, weight, description, color)
         {
             this.sellValue = sellValue;
